pyproject.toml
[build-system]
requires = ["hatchling>=1.27.0"]
build-backend = "hatchling.build"

[project]
name = "juncture"
version = "0.1.0a0"
description = "Multi-backend SQL + Python transformation engine. Local-first, DuckDB-native, Keboola-compatible."
readme = "README.md"
requires-python = ">=3.11"
license = { text = "Apache-2.0" }
authors = [
    { name = "Keboola", email = "devel@keboola.com" },
]
keywords = ["etl", "elt", "data-engineering", "duckdb", "sql", "transformations", "dbt", "keboola"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Database",
    "Topic :: Software Development :: Libraries",
]

dependencies = [
    "duckdb>=1.1.0",
    "sqlglot>=25.0.0",
    "pydantic>=2.7.0",
    "typer>=0.12.0",
    "rich>=13.7.0",
    "pyyaml>=6.0.0",
    "jinja2>=3.1.0",
    "networkx>=3.2.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.27.0",
]

[project.optional-dependencies]
snowflake = [
    "snowflake-connector-python>=3.10.0",
]
bigquery = [
    "google-cloud-bigquery>=3.20.0",
    "db-dtypes>=1.2.0",
]
postgres = [
    "psycopg[binary]>=3.1.0",
]
pandas = [
    "pandas>=2.2.0",
    "pyarrow>=15.0.0",
]
polars = [
    "polars>=1.0.0",
]
lineage = [
    "openlineage-python>=1.18.0",
]
all = [
    "juncture[snowflake,bigquery,postgres,pandas,polars,lineage]",
]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=5.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.6.0",
    "mypy>=1.10.0",
    "pre-commit>=3.7.0",
    "pandas>=2.2.0",
    "pyarrow>=15.0.0",
]

[project.urls]
Homepage = "https://github.com/keboola/new-transformations"
Documentation = "https://github.com/keboola/new-transformations/tree/main/docs"
Repository = "https://github.com/keboola/new-transformations"

[project.scripts]
juncture = "juncture.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["src/juncture"]

[tool.hatch.build.targets.sdist]
include = [
    "src/juncture",
    "docs",
    "examples",
    "skills",
    "README.md",
    "LICENSE",
    "pyproject.toml",
]

[tool.ruff]
line-length = 110
target-version = "py311"
src = ["src", "tests"]

[tool.ruff.lint]
select = [
    "E",
    "W",
    "F",
    "I",
    "B",
    "C4",
    "UP",
    "RUF",
    "SIM",
]
ignore = [
    "E501",
    "B008",
]

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = ["S101"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.pytest.ini_options]
minversion = "8.0"
addopts = ["-ra", "--strict-markers", "--strict-config"]
testpaths = ["tests"]
pythonpath = ["src"]
filterwarnings = [
    "error",
    "ignore::DeprecationWarning:duckdb.*",
    "ignore::DeprecationWarning:sqlglotrs.*",
    "ignore::UserWarning:sqlglot.*",
]

[tool.mypy]
python_version = "3.11"
strict = true
warn_unused_ignores = true
warn_unreachable = true

[[tool.mypy.overrides]]
module = ["duckdb.*", "sqlglot.*", "networkx.*"]
ignore_missing_imports = true
